pyproject.toml
[tool.poetry]
name = "applied-py-2-telegram-bot"
version = "0.1.0"
description = ""
authors = ["Ilia Ilin <ilyinily@gmail.com>"]
readme = "README.md"
packages = [{include = "applied_py_2_telegram_bot", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
aiogram = "^3.17.0"
python-dotenv = "^1.0.1"
fastapi = "^0.115.6"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
